+++ frontend/src/components/index.ts
@@ -23,4 +23,4 @@ export * from 'components/EmptyContent';
 export * from 'components/SnackBar';
 export * from 'components/LoadingSpinner';
 export * from 'components/PrivateOutlet';
-export * from 'components/PublicOutlet';
+export * from 'components/LandingNavigation';
